Reorganize state source files
