+++ src/state/reducers.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
-import uiReducer from './ui_reducer';
-import chordSheetReducer from './chord_sheet_reducer';
+import uiReducer from './ui/reducer';
+import chordSheetReducer from './chord_sheet/reducer';
 
 const rootReducer = combineReducers({
   ui: uiReducer,
